+++ src/plugins/nonebot_bison/platform/bilibili.py
@@ -1,4 +1,5 @@
 import json
+import re
 from typing import Any, Optional
 
 from ..post import Post
@@ -25,6 +26,7 @@ class Bilibili(NewMessage):
     schedule_kw = {"seconds": 10}
     name = "B站"
     has_target = True
+    parse_target_promot = "请输入用户主页的链接"
 
     async def get_target_name(self, target: Target) -> Optional[str]:
         async with http_client() as client:
@@ -35,6 +37,16 @@ class Bilibili(NewMessage):
             if res_data["code"]:
                 return None
             return res_data["data"]["name"]
+
+    async def parse_target(self, target_text: str) -> Target:
+        if re.match(r"\d+", target_text):
+            return Target(target_text)
+        elif match := re.match(
+            r"(?:https?://)?space\.bilibili\.com/(\d+)", target_text
+        ):
+            return Target(match.group(1))
+        else:
+            raise self.ParseTargetException()
 
     async def get_sub_list(self, target: Target) -> list[RawPost]:
         async with http_client() as client: